Move  dashboard view into a folder and split into Vue and JS file to keep consistency
